that function need a pointer to an unsigned int, not a size_t which might be bigger on amd64. also Bug 24 and gentoo bug 162427.

+++ src/ifd/process.c
@@ -423,7 +423,7 @@ int do_memory_write(ifd_reader_t * reader, int unit, ct_tlv_parser_t * args,
 		    ct_tlv_builder_t * resp)
 {
 	unsigned char *data;
-	size_t data_len;
+	unsigned int data_len;
 	unsigned int address;
 	int rc;
